recipe/ftp/requirements.txt
pyftpdlib==1.5.3
PyJWT==1.6.1
